Render proper loop and types
