+++ src/components/AudioGraph/nodes/Gain.tsx
@@ -3,12 +3,12 @@ import { jsx } from "@emotion/core"
 import { useMemo, useEffect, Fragment } from "react"
 import { useSelector, useDispatch } from "react-redux"
 import { NodeComponentProps } from "react-flow-renderer"
-import { setGain, delGain, selectGain, Gain } from "../../../features/activeSound/activeSoundSlice"
+import { setGain, delGain, selectGain } from "../../../features/activeSound/activeSoundSlice"
+import { H1, FormWrapper, NodeBody, Hr } from "../elems/styled"
 import useAudioNodeDefs from "../../../hooks/useAudioNodeDefs"
 import { selectEditMode } from "../../../features/ux/uxSlice"
 import AudioParamDefaults from "../elems/AudioParamDefaults"
-import { AudioParamSetting } from "../elems/AudioParamForm"
-import { H1, FormWrapper, NodeBody, Hr } from "../elems/styled"
+import { Gain, AudioParamSetting } from "../../../audio"
 import AudioParamsView from "../elems/AudioParamsView"
 import HandleOutputs from "../elems/HandleOutputs"
 import HandleInputs from "../elems/HandleInputs"
