+++ src/scripts/audio.ts
@@ -1,8 +1,6 @@
 import Sound from "./Sound"
 import { validateSound } from "./helpers"
-import { SynthLocalStore } from "../components/SelectSound/SelectSound"
-
-export const AUDIO_CONTEXT_DESTINATION = "destination"
+import { SynthStore } from "../audio"
 
 // @ts-ignore
 // eslint-disable-next-line no-native-reassign
@@ -20,7 +18,7 @@ export const restartAudioContext = async () => {
 export const loadSound = (name: string, ctx?: BaseAudioContext) => {
   const data = localStorage.getItem(name)
   if (!data) return null
-  const obj: SynthLocalStore = JSON.parse(data)
+  const obj: SynthStore = JSON.parse(data)
   if (!validateSound(obj)) return null
 
   const s = new Sound(ctx ?? audioContext)
